fix: merge upsert named tests to satisfy pkg-file-count limit
